removed unused test draft. Moved interpolate tests from configfile to config
